+++ imouapi/device.py
@@ -1,10 +1,9 @@
 """High level API to discover and interacting with Imou devices and their sensors."""
 import logging
-
-import aiohttp
+from typing import Union
 
 from imouapi.api import ImouAPIClient
-from imouapi.const import BINARY_SENSORS, IMOU_CAPABILITIES, IMOU_SWITCHES, SENSORS, SUPPORTED_SWITCHES
+from imouapi.const import BINARY_SENSORS, IMOU_CAPABILITIES, IMOU_SWITCHES, SENSORS
 from imouapi.device_entity import ImouBinarySensor, ImouEntity, ImouSensor, ImouSwitch
 from imouapi.exceptions import InvalidResponse
 
@@ -16,31 +15,19 @@ class ImouDevice:
 
     def __init__(
         self,
-        app_id: str,
-        app_secret: str,
+        api_client: ImouAPIClient,
         device_id: str,
-        websession: aiohttp.ClientSession,
-        base_url: str = None,
-        timeout: int = None,
     ) -> None:
         """
         Initialize the instance.
 
         Parameters:
-            app_id: appID from https://open.imoulife.com/consoleNew/myApp/appInfo
-            app_secret: appID from https://open.imoulife.com/consoleNew/myApp/appInfo
+            api_client: an ImouAPIClient instance
             device_id: device id
-            websession: aiohttp client session
-            base_url: API base url (optional)
-            timeout: request timeout in seconds (optional)
         """
-        # initialize the properties
-        self._app_id = app_id
-        self._app_secret = app_secret
-        self._websession = websession
-        self._timeout = timeout
-        self._base_url = base_url
+        self._api_client = api_client
         self._device_id = device_id
+
         self._catalog = "N.A."
         self._firmware = "N.A."
         self._name = "N.A."
@@ -51,12 +38,8 @@ class ImouDevice:
         self._capabilities: list[str] = []
         self._switches: list[str] = []
         self._sensor_instances: dict[str, list] = {"switch": [], "sensor": [], "binary_sensor": []}
-        self._last_alarm = "N.A."
-        # setup the API client
-        self.api_client = ImouAPIClient(app_id, app_secret, websession, base_url, timeout)
-        # other status variables
+
         self._initialized = False
-        self._connected = False
         self._enabled = True
 
     def get_device_id(self) -> str:
@@ -89,29 +72,46 @@ class ImouDevice:
         """Get online."""
         return self._online
 
-    def get_sensors(self, platform: str) -> list[ImouEntity]:
-        """Get sensor instances."""
+    def get_all_sensors(self) -> list[ImouEntity]:
+        """Get all the sensor instances."""
+        sensors = []
+        for (
+            platform,  # pylint: disable=unused-variable
+            sensor_instances_array,
+        ) in self._sensor_instances.items():
+            for sensor_instance in sensor_instances_array:
+                sensors.append(sensor_instance)
+        return sensors
+
+    def get_sensors_by_platform(self, platform: str) -> list[ImouEntity]:
+        """Get sensor instances associated to a given platform."""
+        if platform not in self._sensor_instances:
+            return []
         return self._sensor_instances[platform]
+
+    def get_sensor_by_name(self, name: str) -> Union[ImouSensor, ImouBinarySensor, ImouSwitch, None]:
+        """Get sensor instance with a given name."""
+        for (
+            platform,  # pylint: disable=unused-variable
+            sensor_instances_array,
+        ) in self._sensor_instances.items():
+            for sensor_instance in sensor_instances_array:
+                if sensor_instance.get_name() == name:
+                    return sensor_instance
+        return None
 
     def set_enabled(self, value: bool) -> None:
         """Set enable."""
         self._enabled = value
 
-    def get_last_alarm(self) -> str:
-        """Get last alarm."""
-        return self._last_alarm
-
-    async def async_connect(self) -> bool:
-        """Connect to the API."""
-        status = await self.api_client.async_connect()
-        if status:
-            self._connected = True
-        return status
+    def is_enabled(self) -> bool:
+        """Is enabled."""
+        return self._enabled
 
     async def async_initialize(self) -> None:
-        """Initialize the instance by retrieving the device details."""
+        """Initialize the instance by retrieving the device details and associated sensors."""
         # get the details for this device from the API
-        device_array = await self.api_client.async_api_deviceBaseDetailList([self._device_id])
+        device_array = await self._api_client.async_api_deviceBaseDetailList([self._device_id])
         if "deviceList" not in device_array or len(device_array["deviceList"]) != 1:
             raise InvalidResponse(f"deviceList not found in {str(device_array)}")
         # reponse is an array, our data is in the first element
@@ -135,20 +135,19 @@ class ImouDevice:
                 for capability in self._capabilities:
                     if switch_type.lower() == capability.lower():
                         self._switches.append(switch_type)
-                        # if the switch is supported, create an instance and save it
-                        if switch_type in SUPPORTED_SWITCHES:
-                            switch_instance = ImouSwitch(
-                                self.api_client,
-                                self._device_id,
-                                self.get_name(),
-                                switch_type,
-                            )
-                            self._sensor_instances["switch"].append(switch_instance)
+                        # create an instance and save it
+                        switch_instance = ImouSwitch(
+                            self._api_client,
+                            self._device_id,
+                            self.get_name(),
+                            switch_type,
+                        )
+                        self._sensor_instances["switch"].append(switch_instance)
                         break
             # add lastAlarm sensor
             self._sensor_instances["sensor"].append(
                 ImouSensor(
-                    self.api_client,
+                    self._api_client,
                     self._device_id,
                     self.get_name(),
                     "lastAlarm",
@@ -157,7 +156,7 @@ class ImouDevice:
             # add online binary sensor
             self._sensor_instances["binary_sensor"].append(
                 ImouBinarySensor(
-                    self.api_client,
+                    self._api_client,
                     self._device_id,
                     self.get_name(),
                     "online",
@@ -174,15 +173,13 @@ class ImouDevice:
         """Update device properties and its sensors."""
         if not self._enabled:
             return False
-        if not self._connected:
-            await self.async_connect()
         if not self._initialized:
-            # get the details of the devices
+            # get the details of the device first
             await self.async_initialize()
 
         # check if the device is online
         _LOGGER.debug("[%s] update requested", self.get_name())
-        data = await self.api_client.async_api_deviceOnline(self._device_id)
+        data = await self._api_client.async_api_deviceOnline(self._device_id)
         self._online = data["onLine"] == "1"
 
         # update the status of all the sensors (if the device is online)
@@ -194,15 +191,6 @@ class ImouDevice:
                 for sensor_instance in sensor_instances_array:
                     await sensor_instance.async_update()
         return True
-
-    def enable_sensors(self, value: bool):
-        """Enable/Disable all the registered sensors."""
-        for (
-            platform,  # pylint: disable=unused-variable
-            sensor_instances_array,
-        ) in self._sensor_instances.items():
-            for sensor_instance in sensor_instances_array:
-                sensor_instance.set_enabled(value)
 
     def to_string(self) -> str:
         """Return the object as a string."""
@@ -225,21 +213,13 @@ class ImouDevice:
                 f"{IMOU_CAPABILITIES[capability]} ({capability})" if capability in IMOU_CAPABILITIES else capability
             )
             dump = dump + f"        - {description}\n"
-        dump = dump + "    Available Switches: \n"
-        for sensor_name in self._switches:
-            description = (
-                f"{IMOU_SWITCHES[sensor_name]} ({sensor_name})" if sensor_name in IMOU_SWITCHES else sensor_name
-            )
-            dump = dump + f"        - {description}\n"
-        dump = dump + "    Configured Switches: \n"
+        dump = dump + "    Switches: \n"
         for sensor_instance in self._sensor_instances["switch"]:
             sensor_name = sensor_instance.get_name()
             description = (
                 f"{IMOU_SWITCHES[sensor_name]} ({sensor_name})" if sensor_name in IMOU_SWITCHES else sensor_name
             )
-            is_on = sensor_instance.is_on()
-            status = "ON" if is_on else "OFF"
-            dump = dump + f"        - {description}: {status}\n"
+            dump = dump + f"        - {description}: {sensor_instance.is_on()}\n"
         dump = dump + "    Sensors: \n"
         for sensor_instance in self._sensor_instances["sensor"]:
             sensor_name = sensor_instance.get_name()
@@ -249,57 +229,27 @@ class ImouDevice:
         for sensor_instance in self._sensor_instances["binary_sensor"]:
             sensor_name = sensor_instance.get_name()
             description = f"{BINARY_SENSORS[sensor_name]} ({sensor_name})"
-            is_on = sensor_instance.is_on()
-            status = "ON" if is_on else "OFF"
-            dump = dump + f"        - {description}: {status}\n"
+            dump = dump + f"        - {description}: {sensor_instance.is_on()}\n"
         return dump
 
 
 class ImouDiscoverService:
     """Class for discovering IMOU devices."""
 
-    def __init__(
-        self,
-        app_id: str,
-        app_secret: str,
-        websession: aiohttp.ClientSession,
-        base_url: str = None,
-        timeout: int = None,
-    ) -> None:
+    def __init__(self, api_client: ImouAPIClient) -> None:
         """
         Initialize the instance.
 
         Parameters:
-            app_id: appID from https://open.imoulife.com/consoleNew/myApp/appInfo
-            app_secret: appID from https://open.imoulife.com/consoleNew/myApp/appInfo
-            websession: aiohttp client session
-            base_url: API base url (optional)
-            timeout: request timeout in seconds (optional)
+            api_client: an ImouAPIClient instance
         """
-        # initialize the properties
-        self._app_id = app_id
-        self._app_secret = app_secret
-        self._websession = websession
-        self._base_url = base_url
-        self._timeout = timeout
-        self._connected = False
-        # setup the API client
-        self.api_client = ImouAPIClient(app_id, app_secret, websession, base_url, timeout)
-
-    async def async_connect(self) -> bool:
-        """Connect to the API."""
-        status = await self.api_client.async_connect()
-        if status:
-            self._connected = True
-        return status
+        self._api_client = api_client
 
     async def async_discover_devices(self) -> dict:
         """Discover registered devices and return a dict device name -> device object."""
-        if not self._connected:
-            await self.async_connect()
         _LOGGER.debug("Starting discovery")
         # get the list of devices
-        devices_data = await self.api_client.async_api_deviceBaseList()
+        devices_data = await self._api_client.async_api_deviceBaseList()
         if "deviceList" not in devices_data or "count" not in devices_data:
             raise InvalidResponse(f"deviceList or count not found in {devices_data}")
         _LOGGER.debug("Discovered %d registered devices", devices_data["count"])
@@ -307,13 +257,7 @@ class ImouDiscoverService:
         devices = {}
         for device_data in devices_data["deviceList"]:
             # create a a device instance from the device id and initialize it
-            device = ImouDevice(
-                self._app_id,
-                self._app_secret,
-                device_data["deviceId"],
-                self._websession,
-            )
-            await device.async_connect()
+            device = ImouDevice(self._api_client, device_data["deviceId"])
             await device.async_initialize()
             _LOGGER.debug("   - %s", device.to_string())
             devices[f"{device.get_name()}"] = device